requirements.txt
Flask==2.3.2
tensorflow-cpu==2.20.0
opencv-python==4.7.0.72
numpy==1.26.4
pandas==2.1.0
scikit-learn==1.7.2
gunicorn==20.1.0

